Implemented ammo counts to player HUD and added Scoop Images to the hud
There were platforms in the prototype scene that were bugging out, had to remove those and fix a typing issue for the mint monster
