+++ Assets/Prefabs/UpdatedGameManager2.0.prefab.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: 77dbf2c03f79413439d0fe3c5fe9d783
+PrefabImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
