+++ Assets/Scripts/ScoopHUDManager.cs.meta
