+++ Assets/Scripts/UI/AmmoByScoopType.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 078295cf2027a224fa8cf7a767e91377
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
